refactor: consolidate into wikipedia_global_message_cleanup package
Rename src/lib/ to src/wikipedia_global_message_cleanup/ and move
src/check_last_contribution.py into the package as cli.py. Update
all imports to use the new package name (relative imports in cli.py,
absolute in tests). Simplify pyproject.toml packaging config.
