+++ pyproject.toml
@@ -20,10 +20,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/lib", "src/check_last_contribution.py"]
-
-[tool.hatch.build.targets.wheel.sources]
-src = ""
+packages = ["src/wikipedia_global_message_cleanup"]
 
 [project.scripts]
-check-last-contribution = "check_last_contribution:main"
+check-last-contribution = "wikipedia_global_message_cleanup.cli:main"
