+++ src/wikipedia_global_message_cleanup/analyzer.py
